[ALPS09944626] DISP: enlarge mtcmos debounce size for dpc init
[Description]
enlarge mtcmos debounce size for dpc init

[Test]
boot

MTK-Commit-Id: 33663cc004cfaa39b49f7307217b8e5c6e91d570

CR-Id: ALPS09944626
Feature: V-Idle

+++ drivers/gpu/drm/mediatek/mediatek_v2/mtk_disp_vidle.h
@@ -53,7 +53,7 @@ enum mtk_vidle_hint_type {
 };
 
 struct mtk_vidle_hint {
-	u8 mtcmos_debounce;		/* 0: okay */
+	u32 mtcmos_debounce;		/* 0: okay */
 	u8 mode_switch_debounce;	/* 0: okay */
 	u8 doze_debounce;		/* 0: okay */
 	u8 crtc_fuse;			/* 0: okay, or +1 by others(multi crtc, DP, ...) */
